Validações em telas para impedir exclusão de dependencias finalizadas. Apenas tela de pessoa ainda com problemas.

+++ src/br/com/erp/resources/PessoaImp.java
@@ -11,10 +11,12 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
 import br.com.erp.json.ParamJson;
+import br.com.erp.model.CapCar;
 import br.com.erp.model.Contato;
 import br.com.erp.model.Endereco;
 import br.com.erp.model.Pessoa;
 import br.com.erp.model.Produto;
+import br.com.erp.model.TipoContato;
 import br.com.erp.util.UnidadePersistencia;
 
 @Path("pessoa")
@@ -119,21 +121,53 @@ public class PessoaImp {
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public void remove(ParamJson paramJson) {
 		EntityManager em = UnidadePersistencia.createEntityManager();
-
+		
 		try {
 			Pessoa pessoa = em.find(Pessoa.class, paramJson.getInt1());
 			em.getTransaction().begin();
+			CapCar capCar = obterDependencia(paramJson.getInt1());
+			
+			if(capCar != null) {
+				throw new RuntimeException("Existe uma dependência relacionada a este registro: " + capCar.getId());
+			}
 			em.remove(pessoa);
 			em.getTransaction().commit();
 
 		} catch (Exception e) {
 			e.printStackTrace();
-			em.getTransaction().rollback();
-
+			if(em.getTransaction().isActive()){
+				em.getTransaction().rollback();
+			}
+			throw e;
 		} finally {
 			em.close();
 		}
 
+	}
+	
+	public CapCar obterDependencia(Integer id) {
+		EntityManager em = UnidadePersistencia.createEntityManager();
+		
+		CapCar capCar = null;
+		
+		try {
+			capCar = (CapCar) em.createQuery("select a "
+											   +"from capcar a "
+											   +"where a.cliente.id = :id"
+											   +"and tipo a = 'P'"
+											   +"or tipo a = 'R'"
+											   +"and status a = 0"
+											   +"or status a = 1")
+								  .setParameter("id", id)
+								  .setMaxResults(1)
+					              .getSingleResult();
+		}catch(Exception e){
+			e.printStackTrace();
+		}finally {
+			em.close();
+		}
+		
+		return capCar;
 	}
 
 }
